+++ tests/jobs/test_scheduler_job.py
@@ -44,11 +44,13 @@ from airflow.utils.db import create_session, provide_session
 from airflow.utils.state import State
 from tests.compat import MagicMock, Mock, PropertyMock, patch
 from tests.compat import mock
-from tests.test_core import TEST_DAG_FOLDER
+# from tests.test_core import TEST_DAG_FOLDER
 from tests.executors.test_executor import TestExecutor
 from tests.test_utils.db import clear_db_dags, clear_db_errors, clear_db_pools, \
     clear_db_runs, clear_db_sla_miss, set_default_pool_slots
 
+TEST_DAG_FOLDER = os.path.join(
+    os.path.dirname(os.path.realpath('../test_core.py')), 'dags')
 
 DEFAULT_DATE = timezone.datetime(2016, 1, 1)
 TRY_NUMBER = 1
