+++ tests/sensors/test_hdfs_sensor.py
@@ -23,7 +23,8 @@ from datetime import timedelta
 from airflow.exceptions import AirflowSensorTimeout
 from airflow.sensors.hdfs_sensor import HdfsSensor
 from airflow.utils.timezone import datetime
-from tests.test_core import FakeHDFSHook
+from tests.test_utils.fake_hdfs_hook import FakeHDFSHook
+
 
 DEFAULT_DATE = datetime(2015, 1, 1)
 TEST_DAG_ID = 'unit_test_dag'
