+++ tests/task/task_runner/test_standard_task_runner.py
@@ -31,7 +31,9 @@ from airflow.utils.state import State
 
 from logging.config import dictConfig
 
-from tests.test_core import TEST_DAG_FOLDER
+# from tests.test_core import TEST_DAG_FOLDER
+TEST_DAG_FOLDER = os.path.join(
+    os.path.dirname(os.path.realpath('../../test_core.py')), 'dags')
 
 DEFAULT_DATE = timezone.datetime(2016, 1, 1)
